Extract bounty API routes into app/bounty_api.py
Moves bounty listing, CRUD, payment, close, and reconciliation route
registration out of the main application module into a focused
app/bounty_api module.

What moved:
- list_bounties_by_status search/filter logic
- GET /api/v1/bounties and /api/v1/bounties/summary
- GET /api/v1/bounties/{bounty_id}
- POST /api/v1/bounties (create)
- POST /api/v1/bounties/{bounty_id}/pay
- POST /api/v1/bounties/{bounty_id}/close
- GET /api/v1/reconciliation/payouts
- GET /api/v1/admin/webhook-events
- Payout response shaping helpers

main.py drops from 1427 to 1196 lines. The bounty page routes in
main.py now call through the returned functions from
register_bounty_api_routes, keeping the same public behavior.

13 new route-level tests cover status filtering, query search,
auth gating on admin-only endpoints, and error responses.

Checks: 384 passed, ruff format/check clean, mypy clean.
